# Exponents optimization