# 📄 Description

Performance improvement on the `full_exponents` array generation. The previous list comprehension algorithm is replaced by a generator-based algorithm. The new approach offers significant speedups, especially for polynomials with high degrees or many variables.

# 🎯 Objectives

1. Implement the generator-based algorithm
2. Integration of the new method in the base code
3. Comparison of performance between the methods

# ✅ Tasks

- [x] Implementation
- [x] Refactor
- [ ] Benchmark