@ximiraxelo 2025-07-28
## Benchmark

The benchmark was conducted by considering a grid of `n_vars` values and another of  `degree` values. Each method was tested 100 times across 7 repetitions, using the builtin [`timeit.repeat()`](https://docs.python.org/3/library/timeit.html#timeit.repeat) function. The mean execution time and standard deviation were computed for each configuration.

### 🆕 New method (Generator)

| n_vars 	| degree 	| avg_time               	| std_dev                	|
|--------	|--------	|------------------------	|------------------------	|
| 1      	| 2      	| 5.343142858172152e-06  	| 2.614449321594997e-06  	|
| 1      	| 4      	| 5.3115714269681896e-06 	| 2.218180388609504e-06  	|
| 1      	| 6      	| 6.064428571335157e-06  	| 3.29904496321511e-06   	|
| 1      	| 8      	| 6.560428572421577e-06  	| 1.8103091190851376e-06 	|
| 2      	| 2      	| 1.5031999998557563e-05 	| 3.3645620822627647e-06 	|
| 2      	| 4      	| 1.6592571429490426e-05 	| 1.516743585765139e-06  	|
| 2      	| 6      	| 2.4216142858546973e-05 	| 3.405188706786326e-06  	|
| 2      	| 8      	| 3.6304571431173826e-05 	| 6.05113162620679e-06   	|
| 3      	| 2      	| 1.6581999998379614e-05 	| 3.885896588327068e-06  	|
| 3      	| 4      	| 3.3975714287797535e-05 	| 2.6917818186262713e-06 	|
| 3      	| 6      	| 7.463599999807358e-05  	| 7.861790053750828e-06  	|
| 3      	| 8      	| 0.00013573685714228693 	| 7.542644318126788e-06  	|
| 4      	| 2      	| 2.879471428579044e-05  	| 9.33455365278465e-06   	|
| 4      	| 4      	| 0.00015989728571444824 	| 0.00011142251689799228 	|
| 4      	| 6      	| 0.00023827500000022286 	| 8.406144420596367e-05  	|
| 4      	| 8      	| 0.0004454481428573282  	| 2.2688688981165103e-05 	|
| 5      	| 2      	| 3.207257143003517e-05  	| 1.8723635059743557e-06 	|
| 5      	| 4      	| 0.00019376957142932434 	| 4.7717884432967886e-05 	|
| 5      	| 6      	| 0.0004880079999981847  	| 2.6997757369905607e-05 	|
| 5      	| 8      	| 0.001357442428570721   	| 0.00021723776297186145 	|
| 6      	| 2      	| 4.5325714285614337e-05 	| 2.9101811299919914e-06 	|
| 6      	| 4      	| 0.00025219828571347797 	| 5.868226616042008e-06  	|
| 6      	| 6      	| 0.0010603498571435921  	| 5.2575680254479716e-05 	|
| 6      	| 8      	| 0.0034630082857139703  	| 0.00041696779143549626 	|
| 7      	| 2      	| 7.128328571395417e-05  	| 1.6086762203708282e-05 	|
| 7      	| 4      	| 0.0005144284285688627  	| 9.321204763341524e-05  	|
| 7      	| 6      	| 0.0022619192857135724  	| 0.00012164171584973567 	|
| 7      	| 8      	| 0.007770518428573529   	| 0.00024673390058986794 	|
| 8      	| 2      	| 8.876857143020192e-05  	| 1.2879860476801403e-05 	|
| 8      	| 4      	| 0.0007421512857139792  	| 3.690438425338009e-05  	|
| 8      	| 6      	| 0.004595030857145892   	| 0.0006797513724469928  	|
| 8      	| 8      	| 0.01748069442857351    	| 0.002418262369214497   	|

### 💀 Old method (List comprehension)

| n_vars 	| degree 	| avg_time               	| std_dev                	|
|--------	|--------	|------------------------	|------------------------	|
| 1      	| 2      	| 2.982571430558372e-06  	| 8.980698125330385e-07  	|
| 1      	| 4      	| 7.472714283462015e-06  	| 1.5619208619021026e-06 	|
| 1      	| 6      	| 4.570142859198053e-06  	| 1.025355329197875e-06  	|
| 1      	| 8      	| 5.093428570554741e-06  	| 8.614447771764554e-07  	|
| 2      	| 2      	| 4.197428571777501e-06  	| 4.4235695186650407e-07 	|
| 2      	| 4      	| 9.904285715102951e-06  	| 1.6828257461705685e-06 	|
| 2      	| 6      	| 2.878942857023503e-05  	| 7.407341834520763e-06  	|
| 2      	| 8      	| 2.7903428572244595e-05 	| 8.725122069459725e-06  	|
| 3      	| 2      	| 8.02828571327804e-06   	| 1.58362953124396e-06   	|
| 3      	| 4      	| 2.7304571431159276e-05 	| 3.687202072268816e-06  	|
| 3      	| 6      	| 6.873699999882124e-05  	| 4.479111231496172e-06  	|
| 3      	| 8      	| 0.00013946885714436318 	| 5.269644429275308e-06  	|
| 4      	| 2      	| 1.6376571428346715e-05 	| 3.285482575183491e-07  	|
| 4      	| 4      	| 0.00010970742856963105 	| 7.992831535255898e-06  	|
| 4      	| 6      	| 0.00036655814285560545 	| 5.4009932299408515e-06 	|
| 4      	| 8      	| 0.0009805912857141915  	| 1.420609273191622e-05  	|
| 5      	| 2      	| 4.1665000001428e-05    	| 3.7878305120578194e-06 	|
| 5      	| 4      	| 0.00046051428571445283 	| 7.883242368503147e-06  	|
| 5      	| 6      	| 0.0024683474285692292  	| 0.0001252355649018376  	|
| 5      	| 8      	| 0.008632178142858849   	| 0.000390883808396208   	|
| 6      	| 2      	| 0.00011683957142720049 	| 1.0765482656785717e-05 	|
| 6      	| 4      	| 0.002247082428574296   	| 4.049989784659941e-05  	|
| 6      	| 6      	| 0.016868551571428074   	| 0.0006896842229052243  	|
| 6      	| 8      	| 0.07445273757142656    	| 0.0015881244900375444  	|
| 7      	| 2      	| 0.000339029714285581   	| 5.688317570790643e-06  	|
| 7      	| 4      	| 0.011417636428571182   	| 0.0001856579000140263  	|
| 7      	| 6      	| 0.11790776428571396    	| 0.0006171879732962491  	|
| 7      	| 8      	| 0.7083946798571431     	| 0.030964394379290848   	|
| 8      	| 2      	| 0.0011672554285736883  	| 0.00012129268281070265 	|
| 8      	| 4      	| 0.061305858714284636   	| 0.002750323238977757   	|
| 8      	| 6      	| 0.9284854342857151     	| 0.017621760710947565   	|
| 8      	| 8      	| 7.019319684857144      	| 0.726147548731838      	|

### 📈 Comparison

<img width="2043" height="840" alt="benchmark_plot" src="https://github.com/user-attachments/assets/f24a5b16-a22e-40b1-b859-003b4ee74a17" />